Relocate elasticsearch to elasticsearch5
