+++ MANIFEST.in
@@ -7,7 +7,7 @@ include README.rst
 include README
 include tox.ini
 include setup.py
-include elasticsearch/connection/esthrift/Rest-remote
+include elasticsearch5/connection/esthrift/Rest-remote
 recursive-include docs *
 
 prune docs/_build
