+++ elasticsearch5/compat.py
