+++ elasticsearch5/connection/base.py
