+++ elasticsearch5/connection/pooling.py
