+++ elasticsearch5/connection_pool.py
