+++ elasticsearch5/exceptions.py
