+++ setup.py
@@ -28,7 +28,7 @@ if sys.version_info[:2] == (2, 6):
     install_requires.append('unittest2')
 
 setup(
-    name = 'elasticsearch',
+    name = 'elasticsearch5',
     description = "Python client for Elasticsearch",
     license="Apache License, Version 2.0",
     url = "https://github.com/elastic/elasticsearch-py",
